randint {reformulas}	R Documentation

Remove all random slopes from a formula, while retaining random intercepts.

Description

Remove all random slopes from a formula, while retaining random intercepts.

Usage

randint(form)

Arguments

form
A formula

Value

The new formula

Examples

f <- ~ 1 + a  + b + (a | f) + (1 + a | g) + (a + b | h ) + (1 + a + b | i)
randint(f)

[Package reformulas version 0.4.4 Index]